# aws/services/openapi.yaml
openapi: 3.0.3
info:
  title: CardDemo online services
  version: "1.0.0"
  description: |
    REST replacement for the CardDemo CICS/BMS online programs (contract: `aws/contracts/api.md`).
    Every 4xx/5xx response uses the `ErrorResponse` envelope; `message` reproduces the legacy COBOL
    screen message and `legacyProgram` names the program the rule came from.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /cards:
    get:
      tags: [cards]
      summary: Browse cards (COCRDLIC)
      parameters:
        - { name: acctId, in: query, schema: { type: string, pattern: '^[0-9]{1,11}$' } }
        - { name: cardNum, in: query, schema: { type: string, pattern: '^[0-9]{16}$' } }
        - $ref: '#/components/parameters/StartKey'
        - $ref: '#/components/parameters/Direction'
        - $ref: '#/components/parameters/PageSize7'
      responses:
        "200":
          description: Page of cards
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    properties:
                      items: { type: array, items: { $ref: '#/components/schemas/CardSummary' } }
        "400": { $ref: '#/components/responses/ValidationError' }
        "401": { $ref: '#/components/responses/Unauthorized' }
  /cards/{cardNum}:
    parameters:
      - { name: cardNum, in: path, required: true, schema: { type: string } }
    get:
      tags: [cards]
      summary: Card detail (COCRDSLC)
      parameters:
        - { name: acctId, in: query, schema: { type: string } }
      responses:
        "200":
          description: Card
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CardDetail' }
        "400": { $ref: '#/components/responses/ValidationError' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
    put:
      tags: [cards]
      summary: Update card (COCRDUPC); expiry day is kept from the stored card as on the legacy screen
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CardUpdateRequest' }
      responses:
        "200":
          description: "Changes committed to database"
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CardDetail' }
        "400": { $ref: '#/components/responses/ValidationError' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
        "409": { $ref: '#/components/responses/Conflict' }
        "422": { $ref: '#/components/responses/BusinessRule' }
...
  /users:
    get:
      tags: [users]
      summary: Browse users (COUSR00C) — ADMIN only
      parameters:
        - $ref: '#/components/parameters/StartKey'
        - $ref: '#/components/parameters/Direction'
        - $ref: '#/components/parameters/PageSize10'
      responses:
        "200":
          description: Page of users
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    properties:
                      items: { type: array, items: { $ref: '#/components/schemas/UserSummary' } }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [users]
      summary: Add user (COUSR01C) — ADMIN only
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateUserRequest' }
      responses:
        "201":
          description: "User <id> has been added ..."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserDetail' }
        "400": { $ref: '#/components/responses/ValidationError' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "409": { $ref: '#/components/responses/Conflict' }
  /users/{userId}:
    parameters:
      - { name: userId, in: path, required: true, schema: { type: string, maxLength: 8 } }
    get:
      tags: [users]
      summary: User detail (COUSR02C / COUSR03C lookup) — ADMIN only
      responses:
        "200":
          description: User
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserDetail' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
    put:
      tags: [users]
      summary: Update user (COUSR02C) — ADMIN only
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateUserRequest' }
      responses:
        "200":
          description: "User <id> has been updated ..."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserDetail' }
        "400": { $ref: '#/components/responses/ValidationError' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
        "409": { $ref: '#/components/responses/Conflict' }
        "422": { $ref: '#/components/responses/BusinessRule' }
    delete:
      tags: [users]
      summary: Delete user (COUSR03C) — ADMIN only
      responses:
        "204": { description: Deleted }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    StartKey:
      name: startKey
      in: query
      description: Key of the row adjacent to the requested page (lastKey for next, firstKey for prev)
      schema: { type: string }
    Direction:
      name: direction
      in: query
      schema: { type: string, enum: [next, prev], default: next }
    PageSize7:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 7 }
    PageSize10:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 10 }
  responses:
    ValidationError:
      description: VALIDATION_ERROR / INVALID_REQUEST
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: UNAUTHENTICATED / INVALID_CREDENTIALS
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: FORBIDDEN
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: NOT_FOUND
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: DUPLICATE / CONCURRENT_UPDATE / LOCKED / INTEGRITY_VIOLATION
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    BusinessRule:
      description: BUSINESS_RULE
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [errorCode, message, timestamp]
      properties:
        errorCode:
          type: string
          enum: [VALIDATION_ERROR, INVALID_REQUEST, UNAUTHENTICATED, INVALID_CREDENTIALS, FORBIDDEN, NOT_FOUND,
                 DUPLICATE, CONCURRENT_UPDATE, LOCKED, INTEGRITY_VIOLATION, BUSINESS_RULE, INTERNAL_ERROR]
        message: { type: string, example: "Account ID NOT found..." }
        fieldErrors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
        legacyProgram: { type: string, example: COBIL00C }
        timestamp: { type: string, format: date-time }
    Page:
      type: object
      properties:
        firstKey: { type: string, nullable: true }
        lastKey: { type: string, nullable: true }
        hasNext: { type: boolean }
        hasPrev: { type: boolean }
        message: { type: string, description: Legacy browse message when a boundary is reached }
...
    CardSummary:
      type: object
      properties:
        cardNum: { type: string }
        acctId: { type: integer, format: int64 }
        activeStatus: { type: string }
    CardDetail:
      type: object
      properties:
        cardNum: { type: string }
        acctId: { type: integer, format: int64 }
        cvvCd: { type: integer }
        embossedName: { type: string }
        expirationDate: { type: string, format: date }
        activeStatus: { type: string }
        version: { type: integer, format: int64 }
        message: { type: string }
    CardUpdateRequest:
      type: object
      required: [acctId, version]
      properties:
        acctId: { type: string }
        embossedName: { type: string, maxLength: 50 }
        activeStatus: { type: string, enum: [Y, N] }
        expirationDate: { type: string, description: "yyyy-MM[-dd]; only year and month are editable" }
        version: { type: integer, format: int64 }
...
    UserSummary:
      type: object
      properties:
        userId: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        userType: { type: string, enum: [A, U] }
    UserDetail:
      allOf:
        - $ref: '#/components/schemas/UserSummary'
        - type: object
          properties:
            version: { type: integer, format: int64 }
            message: { type: string }
    CreateUserRequest:
      type: object
      properties:
        userId: { type: string, maxLength: 8 }
        firstName: { type: string, maxLength: 20 }
        lastName: { type: string, maxLength: 20 }
        password: { type: string, maxLength: 8 }
        userType: { type: string, enum: [A, U] }
    UpdateUserRequest:
      type: object
      required: [version]
      properties:
        firstName: { type: string, maxLength: 20 }
        lastName: { type: string, maxLength: 20 }
        password: { type: string, maxLength: 8, description: Optional; unchanged when omitted }
        userType: { type: string, enum: [A, U] }
        version: { type: integer, format: int64 }
...